[DEV-2067] Moved limit request to new module (#1052)
* [DEV-2067] Moved limit request to new module

* [DEV-2067] Review
