+++ src/subdomains/generic/kyc/dto/input/update-limit-request.dto.ts
@@ -1,6 +1,6 @@
-import { IsDate, IsEnum, IsOptional, IsString } from 'class-validator';
 import { Type } from 'class-transformer';
-import { LimitRequestDecision } from '../limit-request.entity';
+import { IsDate, IsEnum, IsOptional, IsString } from 'class-validator';
+import { LimitRequestDecision } from '../../entities/limit-request.entity';
 
 export class UpdateLimitRequestDto {
   @IsOptional()
