+++ src/subdomains/generic/kyc/kyc.module.ts
@@ -6,12 +6,15 @@ import { UserModule } from '../user/user.module';
 import { KycAdminController } from './controllers/kyc-admin.controller';
 import { KycClientController } from './controllers/kyc-client.controller';
 import { KycController } from './controllers/kyc.controller';
+import { LimitRequestController } from './controllers/limit-request.controller';
 import { KycLog } from './entities/kyc-log.entity';
 import { KycStep } from './entities/kyc-step.entity';
+import { LimitRequest } from './entities/limit-request.entity';
 import { NameCheckLog } from './entities/name-check-log.entity';
 import { StepLog } from './entities/step-log.entity';
 import { KycLogRepository } from './repositories/kyc-log.repository';
 import { KycStepRepository } from './repositories/kyc-step.repository';
+import { LimitRequestRepository } from './repositories/limit-request.repository';
 import { NameCheckLogRepository } from './repositories/name-check-log.repository';
 import { StepLogRepository } from './repositories/step-log.repository';
 import { TfaLogRepository } from './repositories/tfa-log.repository';
@@ -23,17 +26,19 @@ import { KycAdminService } from './services/kyc-admin.service';
 import { KycClientService } from './services/kyc-client.service';
 import { KycNotificationService } from './services/kyc-notification.service';
 import { KycService } from './services/kyc.service';
+import { LimitRequestNotificationService } from './services/limit-request-notification.service';
+import { LimitRequestService } from './services/limit-request.service';
 import { NameCheckService } from './services/name-check.service';
 import { TfaService } from './services/tfa.service';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([KycStep, KycLog, NameCheckLog, StepLog]),
+    TypeOrmModule.forFeature([KycStep, KycLog, NameCheckLog, StepLog, LimitRequest]),
     SharedModule,
     NotificationModule,
     forwardRef(() => UserModule),
   ],
-  controllers: [KycController, KycAdminController, KycClientController],
+  controllers: [KycController, KycAdminController, KycClientController, LimitRequestController],
   providers: [
     KycService,
     KycAdminService,
@@ -43,6 +48,7 @@ import { TfaService } from './services/tfa.service';
     NameCheckLogRepository,
     StepLogRepository,
     TfaLogRepository,
+    LimitRequestRepository,
     DilisenseService,
     IdentService,
     FinancialService,
@@ -50,7 +56,9 @@ import { TfaService } from './services/tfa.service';
     KycStepRepository,
     KycNotificationService,
     KycClientService,
+    LimitRequestService,
+    LimitRequestNotificationService,
   ],
-  exports: [DocumentStorageService, NameCheckService, KycAdminService, KycNotificationService],
+  exports: [DocumentStorageService, NameCheckService, KycAdminService, KycNotificationService, LimitRequestService],
 })
 export class KycModule {}
