+++ src/subdomains/generic/kyc/repositories/limit-request.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { BaseRepository } from 'src/shared/repositories/base.repository';
 import { EntityManager } from 'typeorm';
-import { LimitRequest } from './limit-request.entity';
+import { LimitRequest } from '../entities/limit-request.entity';
 
 @Injectable()
 export class LimitRequestRepository extends BaseRepository<LimitRequest> {
