+++ src/subdomains/generic/user/models/kyc/kyc.controller.ts
@@ -9,8 +9,8 @@ import { UserRole } from 'src/shared/auth/user-role.enum';
 import { CountryDtoMapper } from 'src/shared/models/country/dto/country-dto.mapper';
 import { CountryDto } from 'src/shared/models/country/dto/country.dto';
 import { KycFileType } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
-import { LimitRequestDto } from '../limit-request/dto/limit-request.dto';
-import { LimitRequestService } from '../limit-request/limit-request.service';
+import { LimitRequestDto } from '../../../kyc/dto/input/limit-request.dto';
+import { LimitRequestService } from '../../../kyc/services/limit-request.service';
 import { KycDataTransferDto } from './dto/kyc-data-transfer.dto';
 import { KycDataDto } from './dto/kyc-data.dto';
 import { KycDocumentType, KycFileDto } from './dto/kyc-file.dto';
